+++ NEWS.md
@@ -5,6 +5,7 @@ Red Panda C++ Version 0.13.3
   - fix: If project's compiler set is not the same with the default compiler set, auto openned project's file will use wrong compiler set to do syntax check.
   - change: symbols that exactly match are sorted to the front in the code suggestion popup list
   - fix: symbols defind locally should be sorted to the front in the code suggestion popup list
+  - fix: when show function tips, can't correctly calcuate the current position in the function param list
 
 Red Panda C++ Version 0.13.2
   - fix: "delete and exit" button in the environtment / folder option page doesn't work correctly